New/ical rss event update setting

+++ includes/ical/class-calendar-plus-ical-sync.php
@@ -32,6 +32,18 @@ class Calendar_Plus_iCal_Sync {
 	protected $default_status;
 
 	/**
+	 * Update events or keep them original
+	 * @var bool
+	 */
+	protected $keep_updated;
+
+	/**
+	 * Parent feed url
+	 * @var string
+	 */
+	protected $feed_url;
+
+	/**
 	 * Calendar_Plus_iCal_Sync constructor.
 	 *
 	 * @param array $events List of parsed events
@@ -41,14 +53,18 @@ class Calendar_Plus_iCal_Sync {
 		$this->events = $events;
 
 		$args = wp_parse_args( $args, [
-			'author'   => 0,
-			'category' => 0,
-			'status'   => 'publish',
+			'author'        => 0,
+			'category'      => 0,
+			'status'        => 'publish',
+			'keep_updated' => 0,
+			'source'        => '',
 		] );
 
-		$this->event_author = $args['author'] ? $args['author'] : get_current_user_id();
+		$this->event_author   = $args['author'] ? $args['author'] : get_current_user_id();
 		$this->event_category = $args['category'];
 		$this->default_status = $args['status'];
+		$this->keep_updated   = $args['keep_updated'];
+		$this->feed_url       = $args['source'];
 	}
 
 	/**
@@ -89,6 +105,8 @@ class Calendar_Plus_iCal_Sync {
 			'categories'   => [],
 		] );
 
+		$event_data_hash = md5( serialize( $event_data ) );
+
 		if ( 'publish' === $event_data['post_status'] ) {
 			$event_data['post_status'] = $this->default_status;
 		}
@@ -102,15 +120,29 @@ class Calendar_Plus_iCal_Sync {
 			'post_content' => $event_data['post_content'],
 		);
 
+		update_post_meta( $event->ID, '_event_feed_url', $this->feed_url );
+
 		if ( $event ) {
+			if ( ! $this->keep_updated ) {
+				return false;
+			}
+
 			if ( 'trash' === $event_data['post_status'] ) {
 				// The feed event has been deleted, delete locally
 				wp_delete_post( $event->ID, true );
 				return $event->ID;
 			}
 
-			if ( ! $event_data['last_updated'] || $event_data['last_updated'] === $event->get_meta( 'ical_last_updated' ) ) {
-				return false;
+			$old_event_hash = $event->get_meta( 'ical_hash' );
+			if( $old_event_hash ) {
+				if ( $old_event_hash === $event_data_hash ) {
+					return false;
+				}
+			}
+			else {
+				if ( ! $event_data['last_updated'] || $event_data['last_updated'] === $event->get_meta( 'ical_last_updated' ) ) {
+					return false;
+				}
 			}
 
 			$post_args['ID'] = $event->ID;
@@ -121,7 +153,6 @@ class Calendar_Plus_iCal_Sync {
 
 			wp_update_post( $post_args );
 			$post_id = $event->ID;
-
 		} else {
 			if ( 'trash' === $event_data['post_status'] ) {
 				return false;
@@ -161,6 +192,8 @@ class Calendar_Plus_iCal_Sync {
 		if ( $event_data['last_updated'] ) {
 			update_post_meta( $post_id, '_ical_last_updated', $event_data['last_updated'] );
 		}
+
+		update_post_meta( $post_id, '_ical_hash', $event_data_hash );
 
 		do_action('calendarp_ical_sync_update_insert', $post_id, $post_args, $event_data );
